# api.yaml
openapi: 3.0.3
...
info:
  title: Placement Portal Application API
  description: >
    REST/JSON API for the Placement Portal (Flask + SQLite + Redis + Celery).
    All endpoints are under /api. Authentication uses a JWT bearer token
    obtained from /api/login and sent as "Authorization: Bearer <token>".
    Validation failures return 422 with {"errors": {field: message}}.
  version: 2.0.0
servers:
  - url: http://127.0.0.1:5000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Unauthorized:
      description: Missing or invalid JWT
    Forbidden:
      description: Authenticated but wrong role / not permitted
security:
  - bearerAuth: []
paths:
...
  /api/drives:
    get:
      tags: [Common]
      summary: List approved open drives (search/filter)
      description: Approved, not-past-deadline drives from active, non-blacklisted companies (Redis-cached). Any authenticated role.
      parameters:
        - { name: q, in: query, schema: { type: string }, description: Search in drive title / company name }
        - { name: branch, in: query, schema: { type: string }, description: Filter by eligible branch }
        - { name: min_cgpa, in: query, schema: { type: number }, description: Show drives with min_cgpa <= value }
      responses:
        "200": { description: List of drives }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/offer-letter/{application_id}:
    get:
      tags: [Common]
      summary: Download an offer letter PDF
      description: Owning student or the drive's company only. Sent as a file attachment (requires JWT header — fetched as a blob by the SPA).
      parameters:
        - { name: application_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: PDF file }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: No offer letter for this application }
  /api/resume/{student_user_id}:
    get:
      tags: [Common]
      summary: Download a student's resume
      description: The student themself, or a company the student applied to.
      parameters:
        - { name: student_user_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Resume file }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: No resume uploaded }

  # --------------------------------------------------------------- Admin ----
  /api/admin/dashboard:
    get:
      tags: [Admin]
      summary: Admin dashboard statistics (cached)
      description: Counts of students, companies, drives, applications, pending approvals. Redis-cached.
      responses:
        "200": { description: Dashboard stats }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/companies:
    get:
      tags: [Admin]
      summary: List/search companies
      parameters:
        - { name: q, in: query, schema: { type: string }, description: Search by company name/email }
      responses:
        "200": { description: List of companies with standing (approved/blacklisted) }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/students:
    get:
      tags: [Admin]
      summary: List/search students
      parameters:
        - { name: q, in: query, schema: { type: string }, description: Search by student name/email }
      responses:
        "200": { description: List of students with standing }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/drives:
    get:
      tags: [Admin]
      summary: List drives, optionally by status
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [Pending, Approved, Rejected, Closed] } }
      responses:
        "200": { description: List of drives }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/company/{user_id}/approve:
    post:
      tags: [Admin]
      summary: Approve a company registration (emails company)
      description: Sets approved=true; on first approval a notification email is sent to the company via Celery.
      parameters:
        - { name: user_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Company approved }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/company/{user_id}/reject:
    post:
      tags: [Admin]
      summary: Reject a company registration
      parameters:
        - { name: user_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Company rejected }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/drive/{drive_id}/approve:
    post:
      tags: [Admin]
      summary: Approve a placement drive
      parameters:
        - { name: drive_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Drive approved (visible to eligible students) }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/drive/{drive_id}/reject:
    post:
      tags: [Admin]
      summary: Reject a placement drive
      parameters:
        - { name: drive_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Drive rejected }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/user/{user_id}/blacklist:
    post:
      tags: [Admin]
      summary: Blacklist a user (reverts company's approved drives)
      description: Blocks login. For companies, all Approved drives revert to Pending and disappear from student listings.
      parameters:
        - { name: user_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: User blacklisted }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/user/{user_id}/deactivate:
    post:
      tags: [Admin]
      summary: Deactivate a user account
      description: Legacy endpoint (superseded by blacklist in the UI); blocks login via active=false.
      parameters:
        - { name: user_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: User deactivated }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/user/{user_id}/activate:
    post:
      tags: [Admin]
      summary: Reinstate (activate + un-blacklist) a user
      parameters:
        - { name: user_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: User activated }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/applications:
    get:
      tags: [Admin]
      summary: List/search all applications
      description: Optional q performs a case-insensitive partial match across student name, drive title and company name.
      parameters:
        - { name: q, in: query, schema: { type: string } }
      responses:
        "200": { description: List of applications }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/admin/reports:
    get:
      tags: [Admin]
      summary: Aggregate placement report data (cached)
      description: Chart data — applications by status, placements per company, monthly activity. Redis-cached.
      responses:
        "200": { description: Report datasets }
        "403": { $ref: "#/components/responses/Forbidden" }

  # ------------------------------------------------------------- Company ----
...